refactor(core): извлечь ядро aiInsights (paste/generate/review карточек)
